+++ components/src/core/score-card/score-card.component.spec.ts
@@ -1,33 +1,33 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ScoreCardComponent } from './scorecard.component';
-import { ScoreCardModule } from './scorecard.module';
+import { ScoreCardComponent } from './score-card.component';
+import { ScoreCardModule } from './score-card.module';
 import { Component } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
 import { count } from '../../utils/test-utils';
 
 @Component({
     template: `
-        <pxb-scorecard>
+        <pxb-score-card>
             <ng-container action-items>
                 <mat-icon>mail</mat-icon>
                 <mat-icon>cloud</mat-icon>
                 <mat-icon>search</mat-icon>
             </ng-container>
-        </pxb-scorecard>
+        </pxb-score-card>
     `,
 })
 class TestScoreCardActions {}
 
 @Component({
     template: `
-        <pxb-scorecard><div body id="test-content">Content Goes Here</div></pxb-scorecard>
+        <pxb-score-card><div body id="test-content">Content Goes Here</div></pxb-score-card>
     `,
 })
 class TestScoreCardContent {}
 
 @Component({
     template: `
-        <pxb-scorecard><div action-row id="test-action-row">Show Details</div></pxb-scorecard>
+        <pxb-score-card><div action-row id="test-action-row">Show Details</div></pxb-score-card>
     `,
 })
 class TestScoreCardActionRow {}
@@ -53,21 +53,21 @@ describe('ScoreCardComponent', () => {
     it('should display a header', () => {
         component.headerTitle = 'Header Title';
         fixture.detectChanges();
-        const title = fixture.nativeElement.querySelector('.pxb-scorecard-title');
+        const title = fixture.nativeElement.querySelector('.pxb-score-card-title');
         expect(title.innerHTML).toBe('Header Title');
     });
 
     it('should display a subheader', () => {
         component.headerSubtitle = 'Header Subheader';
         fixture.detectChanges();
-        const subtitle = fixture.nativeElement.querySelector('.pxb-scorecard-subtitle');
+        const subtitle = fixture.nativeElement.querySelector('.pxb-score-card-subtitle');
         expect(subtitle.innerHTML).toBe('Header Subheader');
     });
 
     it('should display a third line of text in the header', () => {
         component.headerInfo = 'Header Info';
         fixture.detectChanges();
-        const info = fixture.nativeElement.querySelector('.pxb-scorecard-info');
+        const info = fixture.nativeElement.querySelector('.pxb-score-card-info');
         expect(info.innerHTML).toBe('Header Info');
     });
 
@@ -94,18 +94,18 @@ describe('ScoreCardComponent', () => {
 
     it('should enforce class naming conventions', () => {
         const classList = [
-            '.pxb-scorecard',
-            '.pxb-scorecard-header',
-            '.pxb-scorecard-header-background',
-            '.pxb-scorecard-header-wrapper',
-            '.pxb-scorecard-header-text',
-            '.pxb-scorecard-title',
-            '.pxb-scorecard-subtitle',
-            '.pxb-scorecard-info',
-            '.pxb-scorecard-action-items-wrapper',
-            '.pxb-scorecard-body',
-            '.pxb-scorecard-badge-wrapper',
-            '.pxb-scorecard-action-row-wrapper',
+            '.pxb-score-card',
+            '.pxb-score-card-header',
+            '.pxb-score-card-header-background',
+            '.pxb-score-card-header-wrapper',
+            '.pxb-score-card-header-text',
+            '.pxb-score-card-title',
+            '.pxb-score-card-subtitle',
+            '.pxb-score-card-info',
+            '.pxb-score-card-action-items-wrapper',
+            '.pxb-score-card-body',
+            '.pxb-score-card-badge-wrapper',
+            '.pxb-score-card-action-row-wrapper',
         ];
         for (const className of classList) {
             count(fixture, className);