+++ components/src/core/score-card/score-card.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ScoreCardComponent } from './scorecard.component';
+import { ScoreCardComponent } from './score-card.component';
 import { MatCardModule } from '@angular/material/card';
 import { MatDividerModule } from '@angular/material/divider';