+++ demos/storybook/stories/channel-value/basic-config.stories.ts
@@ -2,8 +2,8 @@ import { text } from '@storybook/addon-knobs';
 
 export const withBasicConfig = (): any => ({
     template: `
-          <pxb-channel-value [value]="value"></pxb-channel-value>
-      `,
+        <pxb-channel-value [value]="value"></pxb-channel-value>
+    `,
     props: {
         value: text('value', '123'),
     },