+++ demos/storybook/stories/hero/with-different-image-types.stories.ts
@@ -1,3 +1,5 @@
 export const withDiffImageTypes = (): any => ({
+    // template imported from:
+    // https://github.com/pxblue/angular-component-library/blob/dev/demos/storybook/stories/hero/different-image-types.component.ts
     template: `<image-types-story></image-types-story>`,
 });