+++ CHANGELOG.md
@@ -1,20 +1,32 @@
 # Change Log
 
+## v2.0.0
+
+-   Add new components for ScoreCard, InfoListItem, and Spacer.
+-   `ChannelValue`'s value attribute now accepts both `string` type and `number` type.
+    _Breaking Changes_
+-   We now switched to use `@pxblue/angular-themes` for our application's default style.
+
 ## v1.3.0
-- Create a storybook demo application
-- Fix bug in ChannelValueComponent where font size input was not being used
+
+-   Create a storybook demo application
+-   Fix bug in ChannelValueComponent where font size input was not being used
 
 ## v1.2.1
-- Adds a new component for EmptyState
-- New index file for simpler import syntax
-    - `import {XXX} from '@pxblue/angular-components'`
+
+-   Adds a new component for EmptyState
+-   New index file for simpler import syntax
+    -   `import {XXX} from '@pxblue/angular-components'`
 
 ## v1.1.0
-Enable support for Angular 7+ 
+
+Enable support for Angular 7+
 
 ## v1.0.0
+
 Angular 7-compatible components
 Minor styling fixes
 
 ## v0.0.1
+
 Initial beta release